+++ packages/data-info/index.js
@@ -1,0 +1,7 @@
+import DataInfo from './src/data-info.vue';
+
+DataInfo.install = function (Vue) {
+  Vue.component(DataInfo.name, DataInfo);
+};
+
+export default DataInfo;